Docs and Type Refactoring
Summary:
Overall: Better header documentation and consistent type usage in function declarations. Switched to using `_t` types whenever possible.

Other minor fixes: `const` where possible to clarify in/out pointers, `extern "C"` where applicable

Test Plan: Compiles, no more references to non `_t` types.

Reviewers: berman5, O1 The Ori, brookea

Reviewed By: berman5, O1 The Ori, brookea

Tags: #zorp

Differential Revision: http://phabricator.purduesigbots.com/D54

+++ include/pros/rtos.h
@@ -21,15 +21,30 @@
 extern "C" {
 #endif
 
+// The highest priority that can be assigned to a task. Beware of deadlock.
 #define TASK_PRIORITY_MAX 16
+
+// The lowest priority that can be assigned to a task.
+// This may cause severe performance problems and is generally not recommended.
 #define TASK_PRIORITY_MIN 1
+
+// The default task priority, which should be used for most tasks.
+// Default tasks such as autonomous() inherit this priority.
 #define TASK_PRIORITY_DEFAULT 8
 
+// The recommended stack size for a new task. This stack size is used for
+// default tasks such as autonomous(). This equates to 32,768 bytes, or 128
+// times the default stack size for a task in PROS 2.
 #define TASK_STACK_DEPTH_DEFAULT 0x2000
+
+// The minimal stack size for a task. This equates to 2048 bytes, or 8 times the
+// default stack size for a task in PROS 2.
 #define TASK_STACK_DEPTH_MIN 0x200
 
+// The maximum number of characters allowed in a task's name.
 #define TASK_NAME_MAX_LEN 32
 
+// The maximum timeout value that can be given to, for instance, a mutex grab.
 #define TIMEOUT_MAX ((uint32_t)0xffffffffUL)
 
 typedef void* task_t;
@@ -61,34 +76,37 @@ typedef void* mutex_t;
 
 /**
  * Returns the number of milliseconds since PROS initialized.
+ *
+ * \return The number of milliseconds since PROS initialized
  */
 uint32_t millis(void);
 
 /**
  * Create a new task and add it to the list of tasks that are ready to run.
  *
- * \param out_handle
- *          Used to pass back a handle by which the newly created task can be
- *          referenced.
  * \param function
- *          Pointer to the task entry function
+ *        Pointer to the task entry function
  * \param parameters
- *          Pointer to memory that will be used as a parameter for the task being
- *          created. This memory should not typically come from stack, but rather
- *          from dynamically (i.e., malloc'd) or statically allocated memory.
- * \param name
- *          A descriptive name for the task.  This is mainly used to facilitate
- *          debugging. The name may be up to 32 characters long.
+ *        Pointer to memory that will be used as a parameter for the task being
+ *        created. This memory should not typically come from stack, but rather
+ *        from dynamically (i.e., malloc'd) or statically allocated memory.
  * \param prio
- *          The priority at which the task should run.
- *          TASK_PRIO_DEFAULT plus/minus 1 or 2 is typically used.
+ *        The priority at which the task should run.
+ *        TASK_PRIO_DEFAULT plus/minus 1 or 2 is typically used.
  * \param stack_depth
- *          The number of words (i.e. 4 * stack_depth) available on the task's
- *          stack. TASK_STACK_DEPTH_DEFAULT is typically sufficienct.
- * \return
- *          Will return a handle by which the newly created task can be
- *          referenced. If an error occurred, NULL will be returned and errno
- *          can be checked for hints as to why task_create failed.
+ *        The number of words (i.e. 4 * stack_depth) available on the task's
+ *        stack. TASK_STACK_DEPTH_DEFAULT is typically sufficienct.
+ * \param name
+ *        A descriptive name for the task.  This is mainly used to facilitate
+ *        debugging. The name may be up to 32 characters long.
+ *
+ * This function uses the following values of errno when an error state is
+ * reached:
+ * ENOMEM - The stack cannot be used as the TCB was not created.
+ *
+ * \return A handle by which the newly created task can be
+ *         referenced. If an error occurred, NULL will be returned and errno
+ *         can be checked for hints as to why task_create failed.
  */
 task_t task_create(task_fn_t function, void* parameters, uint8_t prio, uint16_t stack_depth, const char* name);
 
@@ -100,8 +118,8 @@ task_t task_create(task_fn_t function, void* parameters, uint8_t prio, uint16_t 
  * should be freed before the task is deleted.
  *
  * \param task
- *          The handle of the task to be deleted.  Passing NULL will cause the
- *          calling task to be deleted.
+ *        The handle of the task to be deleted.  Passing NULL will cause the
+ *        calling task to be deleted.
  */
 void task_delete(task_t task);
 
@@ -113,9 +131,9 @@ void task_delete(task_t task);
  * To delay cyclically, use task_delay_until().
  *
  * \param milliseconds
- *          The number of milliseconds to wait (1000 milliseconds per second)
+ *        The number of milliseconds to wait (1000 milliseconds per second)
  */
-void task_delay(const unsigned long milliseconds);
+void task_delay(const uint32_t milliseconds);
 
 #define delay(milliseconds) task_delay(milliseconds)
 
@@ -127,19 +145,19 @@ void task_delay(const unsigned long milliseconds);
  * be updated to reflect the time at which the task will unblock.
  *
  * \param prev_time
- *          A pointer to the location storing the setpoint time
+ *        A pointer to the location storing the setpoint time
  * \param delta
- *          The number of milliseconds to wait (1000 milliseconds per second)
+ *        The number of milliseconds to wait (1000 milliseconds per second)
  */
-void task_delay_until(unsigned long* const prev_time, const unsigned long delta);
+void task_delay_until(uint32_t* const prev_time, const uint32_t delta);
 
 /**
  * Obtains the priority of the specified task.
  *
  * \param task
- *          The task to check
- * \return
- *          The priority of the task
+ *        The task to check
+ *
+ * \return The priority of the task
  */
 uint32_t task_get_priority(task_t task);
 
@@ -151,9 +169,9 @@ uint32_t task_get_priority(task_t task);
  * may occur.
  *
  * \param task
- *          The task to set
+ *        The task to set
  * \param prio
- *          The new priority of the task
+ *        The new priority of the task
  */
 void task_set_priority(task_t task, uint32_t prio);
 
@@ -161,9 +179,9 @@ void task_set_priority(task_t task, uint32_t prio);
  * Obtains the state of the specified task.
  *
  * \param task
- *          The task to check
- * \return
- *          The state of the task
+ *        The task to check
+ *
+ * \return The state of the task
  */
 task_state_e_t task_get_state(task_t task);
 
@@ -171,7 +189,7 @@ task_state_e_t task_get_state(task_t task);
  * Suspends the specified task, making it ineligible to be scheduled.
  *
  * \param task
- *          The task to suspend
+ *        The task to suspend
  */
 void task_suspend(task_t task);
 
@@ -179,7 +197,7 @@ void task_suspend(task_t task);
  * Resumes the specified task, making it eligible to be scheduled.
  *
  * \param task
- *          The task to resume
+ *        The task to resume
  */
 void task_resume(task_t task);
 
@@ -189,8 +207,7 @@ void task_resume(task_t task);
  * reaped by the idle task will also be included in the count. Tasks recently
  * created may take one context switch to be counted.
  *
- * \return
- *          The number of tasks that are currently being managed by the kernel.
+ * \return The number of tasks that are currently being managed by the kernel.
  */
 uint32_t task_get_count(void);
 
@@ -198,9 +215,9 @@ uint32_t task_get_count(void);
  * Obtains the name of the specified task.
  *
  * \param task
- *          The task to check
- * \return
- *          A pointer to the name of the task
+ *        The task to check
+ *
+ * \return A pointer to the name of the task
  */
 char const* task_get_name(task_t task);
 
@@ -210,9 +227,9 @@ char const* task_get_name(task_t task);
  * The operation takes a relatively long time and should be used sparingly.
  *
  * \param name
- *          The name to query
- * \return
- *          A task handle with a matching name, or NULL if none were found.
+ *        The name to query
+ *
+ * \return A task handle with a matching name, or NULL if none were found.
  */
 task_t task_get_by_name(char* name);
 
@@ -222,9 +239,9 @@ task_t task_get_by_name(char* name);
  * See https://pros.cs.purdue.edu/v5/tutorials/notifications for details.
  *
  * \param task
- *			The task to notify
- * \return
- *			Always returns true.
+ *			  The task to notify
+ *
+ * \return Always returns true.
  */
 uint32_t task_notify(task_t task);
 
@@ -236,19 +253,19 @@ uint32_t task_notify(task_t task);
  * See https://pros.cs.purdue.edu/v5/tutorials/notifications for details.
  *
  * \param task
- *          The task to notify
+ *        The task to notify
  * \param value
- *          The value used in performing the action
+ *        The value used in performing the action
  * \param action
- *          An action to optionally perform on the receiving task's notification
- *          value
- * \param   prev_value
- *          A pointer to store the previous value of the target task's notification, may be NULL
- * \return
- *          Dependent on the notification action.
- *          For NOTIFY_ACTION_NO_OWRITE: return 0 if the value could be written
- *          without needing to overwrite, 1 otherwise.
- *          For all other NOTIFY_ACTION values: always return 0
+ *        An action to optionally perform on the receiving task's notification
+ *        value
+ * \param prev_value
+ *        A pointer to store the previous value of the target task's notification, may be NULL
+ *
+ * \return Dependent on the notification action.
+ *         For NOTIFY_ACTION_NO_WRITE: return 0 if the value could be written
+ *         without needing to overwrite, 1 otherwise.
+ *         For all other NOTIFY_ACTION values: always return 0
  */
 uint32_t task_notify_ext(task_t task, uint32_t value, notify_action_e_t action, uint32_t* prev_value);
 
@@ -258,11 +275,14 @@ uint32_t task_notify_ext(task_t task, uint32_t value, notify_action_e_t action, 
  * See https://pros.cs.purdue.edu/v5/tutorials/notifications for details.
  *
  * \param clear_on_exit
- *          If true (1), then the notification value is cleared.
- *          If false (0), then the notification value is decremented.
+ *        If true (1), then the notification value is cleared.
+ *        If false (0), then the notification value is decremented.
  * \param timeout
- *          Specifies the amount of time to be spent waiting for a notification
- *          to occur.
+ *        Specifies the amount of time to be spent waiting for a notification
+ *        to occur.
+ *
+ * \return The value of the task's notification value before it is decremented
+ *         or cleared
  */
 uint32_t task_notify_take(bool clear_on_exit, uint32_t timeout);
 
@@ -272,9 +292,9 @@ uint32_t task_notify_take(bool clear_on_exit, uint32_t timeout);
  * See https://pros.cs.purdue.edu/v5/tutorials/notifications for details.
  *
  * \param task
- *          The task to clear
- * \return
- *          False if there was not a notification waiting, true if there was
+ *        The task to clear
+ *
+ * \return False if there was not a notification waiting, true if there was
  */
 bool task_notify_clear(task_t task);
 
@@ -283,9 +303,9 @@ bool task_notify_clear(task_t task);
  *
  * See https://pros.cs.purdue.edu/v5/tutorials/multitasking#mutexes for details.
  *
- * \return
- *          A handle to a newly created mutex. If an error occurred, NULL will be
- *			returned and errno can be checked for hints as to why mutex_create failed.
+ * \return A handle to a newly created mutex. If an error occurred, NULL will be
+ *			   returned and errno can be checked for hints as to
+ *         why mutex_create failed.
  */
 mutex_t mutex_create(void);
 
@@ -296,14 +316,14 @@ mutex_t mutex_create(void);
  * See https://pros.cs.purdue.edu/v5/tutorials/multitasking#mutexes for details.
  *
  * \param mutex
- *          Mutex to attempt to lock.
+ *        Mutex to attempt to lock.
  * \param timeout
- *          Time to wait before the mutex becomes available. A timeout of 0 can
- *          be used to poll the mutex. TIMEOUT_MAX can be used to block indefinitely.
- * \return
- *          True if the mutex was successfully taken, false otherwise. If false
- *          is returned, then errno is set with a hint about why the the mutex
- *          couldn't be taken.
+ *        Time to wait before the mutex becomes available. A timeout of 0 can
+ *        be used to poll the mutex. TIMEOUT_MAX can be used to block indefinitely.
+ *
+ * \return True if the mutex was successfully taken, false otherwise. If false
+ *         is returned, then errno is set with a hint about why the the mutex
+ *         couldn't be taken.
  */
 bool mutex_take(mutex_t mutex, uint32_t timeout);
 
@@ -313,11 +333,11 @@ bool mutex_take(mutex_t mutex, uint32_t timeout);
  * See https://pros.cs.purdue.edu/v5/tutorials/multitasking#mutexes for details.
  *
  * \param mutex
- *          Mutex to unlock.
- * \return
- *          True if the mutex was successfully returned, false otherwise. If false
- *          is returned, then errno is set with a hint about why the mutex couldn't
- *          be returned.
+ *        Mutex to unlock.
+ *
+ * \return True if the mutex was successfully returned, false otherwise. If false
+ *         is returned, then errno is set with a hint about why the mutex couldn't
+ *         be returned.
  */
 bool mutex_give(mutex_t mutex);
